extension: rename goplugin to goplugin-loader (#222)
- rename the goplugin to goplugin-loader to make the naming more clear.

- The `goplugin` will be used to indicate the standalone goplugin built
with `buildmode=plugin` or the plugin mechanism that provided by the
golang.
- The `goplugin-loader` will be used to indicate a golang dynamic module
extension that used to load standalone goplugin.

---------
