+++ extensions/composer/goplugin-loader/config.go
