+++ extensions/composer/goplugin-loader/goplugin_internal_test.go
