+++ extensions/composer/goplugin-loader/imagefetcher/fetcher.go
