+++ extensions/composer/goplugin-loader/imagefetcher/fetcher_e2e_test.go
