+++ extensions/composer/goplugin-loader/imagefetcher/fetcher_test.go
