+++ extensions/composer/goplugin-loader/imagefetcher/keychain.go
