+++ extensions/composer/plugins.go
@@ -7,17 +7,18 @@
 package composer
 
 import (
+	// Go plugin to loader other composer plugins that be compiled into separate shared libraries.
+	_ "github.com/tetratelabs/built-on-envoy/extensions/composer/goplugin-loader"
+
 	_ "github.com/tetratelabs/built-on-envoy/extensions/composer/azure-content-safety/embedded" // Azure Content Safety plugin.
 	_ "github.com/tetratelabs/built-on-envoy/extensions/composer/bedrock-guardrails/embedded"   // Bedrock Guardrails plugin.
 	_ "github.com/tetratelabs/built-on-envoy/extensions/composer/cedar/embedded"                // Cedar authorization plugin.
-	// Go plugin to loader other composer plugins that be compiled into separate shared libraries.
-	_ "github.com/tetratelabs/built-on-envoy/extensions/composer/example/embedded" // Example built-in plugin.
-	_ "github.com/tetratelabs/built-on-envoy/extensions/composer/goplugin"
-	_ "github.com/tetratelabs/built-on-envoy/extensions/composer/jwe-decrypt/embedded"       // JWE decryption plugin.
-	_ "github.com/tetratelabs/built-on-envoy/extensions/composer/opa/embedded"               // OPA authorization plugin.
-	_ "github.com/tetratelabs/built-on-envoy/extensions/composer/openfga/embedded"           // OpenFGA authorization plugin.
-	_ "github.com/tetratelabs/built-on-envoy/extensions/composer/openapi-validator/embedded" // OpenAPI validator plugin.
-	_ "github.com/tetratelabs/built-on-envoy/extensions/composer/saml/embedded"              // SAML SP plugin.
-	_ "github.com/tetratelabs/built-on-envoy/extensions/composer/token-exchange/embedded"    // OAuth2 Token Exchange plugin.
-	_ "github.com/tetratelabs/built-on-envoy/extensions/composer/waf/embedded"               // WAF plugin.
+	_ "github.com/tetratelabs/built-on-envoy/extensions/composer/example/embedded"              // Example built-in plugin.
+	_ "github.com/tetratelabs/built-on-envoy/extensions/composer/jwe-decrypt/embedded"          // JWE decryption plugin.
+	_ "github.com/tetratelabs/built-on-envoy/extensions/composer/opa/embedded"                  // OPA authorization plugin.
+	_ "github.com/tetratelabs/built-on-envoy/extensions/composer/openfga/embedded"              // OpenFGA authorization plugin.
+	_ "github.com/tetratelabs/built-on-envoy/extensions/composer/openapi-validator/embedded"    // OpenAPI validator plugin.
+	_ "github.com/tetratelabs/built-on-envoy/extensions/composer/saml/embedded"                 // SAML SP plugin.
+	_ "github.com/tetratelabs/built-on-envoy/extensions/composer/token-exchange/embedded"       // OAuth2 Token Exchange plugin.
+	_ "github.com/tetratelabs/built-on-envoy/extensions/composer/waf/embedded"                  // WAF plugin.
 )
