+++ website/src/pages/docs/extension-types.mdx
@@ -53,7 +53,7 @@ body modification, metadata, and local replies.
 
 > 🚧 There are some limitations related to how the Go runtime loads dependencies that affect how different Go
 > extensions can be loaded in the same Envoy process. To overcome these, a
-> [special dynamic module](https://github.com/tetratelabs/built-on-envoy/tree/main/extensions/internal/goplugin)
+> [special dynamic module](https://github.com/tetratelabs/built-on-envoy/tree/main/extensions/composer/goplugin-loader)
 > is used to load the Go extensions at runtime.
 
 **Manifest type:** `go`
